chipName*LSST() runs using the new focal plane coord method
This involved hacking a lot of unit tests which compare the
*LSST() methods with the not LSST-specific CameraUtils methods.
Since the latter do not account for the optical distortion we do
not expect these methods to agree any more.  The hacks allow the
methods to disagree in cases where a source lands on the edge
of a chip (in which case, we would expect that the optical
distortion could cause the source to fall off the edge of
the chip).

Future commits will implement more reasonable tests of these
methods