expand pixel coordinate test to all bands
This required some weird if/else chains.  There are cases
where ignoring the filter-component of the optical model
is actually better either at the center of the focal plane
or at its very edge.  In almost all of these cases, the
new transformation is only wrong by a few pixels.